ud_fs_5_catalog/requirements.txt
sqlalchemy
flask
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
psycopg2

Flask-Login
oauth2client
requests
